requirements.txt
appnope==0.1.0
autopep8==1.3.5
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.1
decorator==4.3.0
Django==2.0.4
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
httpie==0.9.9
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.4.0
Pygments==2.2.0
PyJWT==1.6.1
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
